Split batch 2: shared CSS layer, layout store, and 10 leaf components
terminal.css takes the cross-cutting classes (modal family, panel chrome,
tone chips); the panel layout gets a store with the duplicate-"markets"
id fix + saved-layout migration (tested). Extracted: CheatSheetModal,
NewsMarquee, MacroPanel, Spark, AiReadLine, DragHead, AuthModal (busy-flag
split), CommandPalette (rows no longer compute while closed — was per-tick),
AlertsModal, ToastStack + an alerts store whose check() stays allocation-free
on the WS tick path. Marketing pages rename two generic classes that would
have collided with the new global sheet after a /terminal visit.

Page: 10,511 → ~8,100 lines. Portal tests: 150 → 164.
